chore: move floating card types
